Add Validations

+++ backend/src/users/users.service.ts
@@ -12,7 +12,6 @@ import { UpdateUserDto } from './dto/update-user.dto'
 //Lo primero que tengo que hacer es traer la entidad
 @Injectable()
 export class UsersService {
-
     //creo el constructor -Injecto un repositorio
     //le doy un nombre (userRepository) 
     //y le digo que tipo de dato es (Respository) el cual importo
@@ -20,6 +19,8 @@ export class UsersService {
     //private: para que se instancie automaticamente
     constructor(@InjectRepository(User) private userRepository: Respository<User>) { }
 
+
+    //-----------------------------------------------------
     async createUser(user: CreateUserDto) {
 
         const userFound = await this.userRepository.finOne({
@@ -27,7 +28,8 @@ export class UsersService {
         })
 
         if (userFound) {
-            return new HttpException('User already exist', HttpStatus.CONFLICT)
+            //HttpExceptio('mensaje',numero o enum)
+            return new HttpException('User already exists', HttpStatus.CONFLICT)
         }
 
 
@@ -35,11 +37,16 @@ export class UsersService {
         return this.userRepository.save(newUser)
     }
 
+
+
+    //-----------------------------------------------------
     getUsers() {
         return this.userRepository.find()
     }
 
-    getUser(id: number) {
+
+    //-----------------------------------------------------
+    async getUser(id: number) {
         const userFound = await this.userRepository.finOne({
             where: { id }
         })
@@ -49,16 +56,56 @@ export class UsersService {
         return userFound;
     }
 
+
+
+
+    //-----------------------------------------------------
     async deleteUser(id: number) {
+        const userFound = await this.userRepository.findOne({
+            where: { id }
+        })
+
+        if (!userFound) {
+            return new HttpException('User not found', HttpException.NOT_FOUND)
+        }
+
         return this.userRepository.delete({
             id: id
         })
 
     }
 
-    updateUser(id: number, user: UpdateUserDto) {
-        return this.userRepository.update({ id: id }, user)
+
+    //-----------------------------------------------------
+    async updateUser(id: number, user: UpdateUserDto) {
+        const userFound = await this.userRepository.findOne({ where: { id } })
+
+        if (!userFound) {
+            return new HttpException('User nor found', HttpStatus.NOT_FOUND)
+        }
+        //Esto es lo mismo/son dos opciones:
+        //return this.userRepository.update({ id: id }, user)
+        const userUpdated = Object.assign(userFound, user);
+        return this.userRepository.save(userUpdated)
 
     }
 
 }
+
+
+
+
+
+//-----------------------------------------------------
+//Otra forma de escribir el delelete teniendo en cuenta que el resuoltado que 
+//devuelve el detele va a ser "affected= 0" si se encuentra un resultado a 
+//elimina
+
+// async deleteUser(id:number){
+//    const result = await this.userRepository.delete({id})
+//    if (result.affected === 0){
+//        return new HttpException('User not found', HttpStatus.NOT_FOUND);
+//    }
+//    return result
+//}
+
